requirements.txt
-e ./libs/GraphReasoning
llama_cpp_python==0.3.7
langchain-community==0.3.17
llama-index-embeddings-huggingface==0.5.1
llama-cpp-python==0.3.7
openai==1.63.0
ratelimit==2.2.1
azure-ai-ml==1.25.0